Refactor alert manager out of request form handler for reuse
